+++ src/main.go
@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-const VERSION = "0.9.0"
+const VERSION = "0.10.0"
 
 func main() {
 	flag.Parse()