+++ src/pg_schema_column.go
@@ -203,7 +203,8 @@ func (pgSchemaColumn *PgSchemaColumn) toParquetSchemaField() ParquetSchemaField 
 func (pgSchemaColumn *PgSchemaColumn) parquetPrimitiveValue(value string) interface{} {
 	switch strings.TrimLeft(pgSchemaColumn.UdtName, "_") {
 	case "varchar", "char", "text", "bytea", "jsonb", "json", "tsvector", "numeric", "uuid", "interval",
-		"point", "line", "lseg", "box", "path", "polygon", "circle":
+		"point", "line", "lseg", "box", "path", "polygon", "circle",
+		"cidr", "inet", "macaddr", "macaddr8":
 		return value
 	case "bpchar":
 		trimmedValue := strings.TrimRight(value, " ")
@@ -284,7 +285,8 @@ func (pgSchemaColumn *PgSchemaColumn) parquetPrimitiveValue(value string) interf
 func (pgSchemaColumn *PgSchemaColumn) parquetPrimitiveTypes() (primitiveType string, primitiveConvertedType string) {
 	switch strings.TrimLeft(pgSchemaColumn.UdtName, "_") {
 	case "varchar", "char", "text", "bpchar", "bytea", "interval", "jsonb", "json", "tsvector",
-		"point", "line", "lseg", "box", "path", "polygon", "circle":
+		"point", "line", "lseg", "box", "path", "polygon", "circle",
+		"cidr", "inet", "macaddr", "macaddr8":
 		return "BYTE_ARRAY", "UTF8"
 	case "date":
 		return "INT32", "DATE"
@@ -326,7 +328,8 @@ func (pgSchemaColumn *PgSchemaColumn) parquetPrimitiveTypes() (primitiveType str
 func (pgSchemaColumn *PgSchemaColumn) icebergPrimitiveType() string {
 	switch strings.TrimLeft(pgSchemaColumn.UdtName, "_") {
 	case "varchar", "char", "text", "interval", "jsonb", "json", "bpchar", "tsvector",
-		"point", "line", "lseg", "box", "path", "polygon", "circle":
+		"point", "line", "lseg", "box", "path", "polygon", "circle",
+		"cidr", "inet", "macaddr", "macaddr8":
 		return "string"
 	case "uuid":
 		return "uuid"